# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/request-pairing": {
      "post": {
        "operationId": "AuthController_requestPairingCode",
        "summary": "Request pairing code",
        "description": "Request a pairing code to link WhatsApp account with the application",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing code sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "12345678"
                    },
                    "message": {
                      "type": "string",
                      "example": "Pairing code sent successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - User is already paired"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/verify-pairing": {
      "post": {
        "operationId": "AuthController_verifyPairing",
        "summary": "Verify pairing success",
        "description": "Verify that WhatsApp pairing was successful (called by webhook from connector)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/confirm-pairing": {
      "post": {
        "operationId": "AuthController_confirmPairing",
        "summary": "Confirm pairing completion",
        "description": "Called by frontend when user confirms they have completed the pairing process on their phone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPairingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pairing confirmed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "WhatsApp connection not yet confirmed"
          },
          "401": {
            "description": "Invalid or expired pairing token"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login with phone number",
        "description": "Send OTP code to user WhatsApp for login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP envoyé avec succès"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User must complete pairing before logging in"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "summary": "Verify OTP code",
        "description": "Verify OTP code and complete login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "phoneNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "whatsappProfile": {
                          "type": "object"
                        },
                        "lastLoginAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired OTP"
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current user",
        "description": "Get the currently authenticated user information",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "phoneNumber": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "whatsappProfile": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": ["auth"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Monorepo API",
    "description": "Interractive documentation for the Monorepo API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RequestPairingDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          }
        },
        "required": ["phoneNumber"]
      },
      "VerifyPairingDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          },
          "whatsappProfile": {
            "type": "object",
            "description": "WhatsApp profile information",
            "example": {
              "name": "John Doe",
              "picture": "https://..."
            }
          }
        },
        "required": ["phoneNumber", "whatsappProfile"]
      },
      "ConfirmPairingDto": {
        "type": "object",
        "properties": {
          "pairingToken": {
            "type": "string",
            "description": "The pairing token received when requesting the pairing code",
            "example": "a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6a7b8c9d0e1f2"
          }
        },
        "required": ["pairingToken"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          }
        },
        "required": ["phoneNumber"]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in E.164 format",
            "example": "+15551234567"
          },
          "code": {
            "type": "string",
            "description": "6-digit OTP code",
            "example": "123456"
          }
        },
        "required": ["phoneNumber", "code"]
      },
...
    }
  }
}